Rename endpoints folder to routes
